Tuna shawn branch

+++ src/sample/controllers/WalletsDetailControl.java
@@ -1,24 +1,21 @@
 package sample.controllers;
 
 import com.jfoenix.controls.JFXButton;
-import com.jfoenix.controls.JFXComboBox;
 import javafx.fxml.FXML;
-import javafx.scene.control.Label;
-import javafx.scene.control.TextField;
+import javafx.scene.control.*;
 import javafx.scene.image.Image;
 import javafx.scene.image.ImageView;
-import sample.models.Transactions;
 import sample.models.Wallets;
 
 import java.io.IOException;
-import java.util.ArrayList;
+import java.util.Optional;
 
 public class WalletsDetailControl {
     @FXML
-    private ImageView icon;
-    @FXML private Label currency;
-    @FXML private TextField balance, name_label;
-    @FXML private JFXButton save_button, delete_button,
+    ImageView icon;
+    @FXML Label currency;
+    @FXML TextField balance, name_label;
+    @FXML JFXButton save_button, delete_button,
             wallet_button,
             visa_button,
             paypal_button,
@@ -29,7 +26,6 @@ public class WalletsDetailControl {
     private MainFrameControl mainFrameControl;
     private MyWalletControl myWalletControl;
     private Wallets w;
-    private Wallets selected_wallet;
 
     public WalletsDetailControl(MainFrameControl mainFrameControl, MyWalletControl myWalletControl, Wallets w) {
         this.mainFrameControl = mainFrameControl;
@@ -71,16 +67,24 @@ public class WalletsDetailControl {
             w.setBalance(balance);
             w.setIcon_name(chose_icon);
 
-            myWalletControl.clearDetailPanel();
             myWalletControl.populateWalletList();
             if (mainFrameControl.getSelectedWallet().equals(w))
                 myWalletControl.setSelected_panel(w);
+
+            myWalletControl.showNothingToDisplay();
         } catch (NumberFormatException | IOException ignored){
 
         }
     }
 
     private void setDelete_button() throws IOException {
-        myWalletControl.deleteAWallet(w);
+        Alert alert = new Alert(Alert.AlertType.WARNING, "CONFIRM DELETE THIS WALLET?", ButtonType.YES, ButtonType.CANCEL);
+        Optional<ButtonType> pressed = alert.showAndWait();
+        ButtonType button = pressed.orElse(ButtonType.CANCEL);
+        if (button == ButtonType.YES){
+            myWalletControl.deleteAWallet(w);
+            myWalletControl.showNothingToDisplay();
+        }
+
     }
 }
